M17: HAL portability cut (partial - vmm + syscall deferred)
Introduces kernel/includes/hal_api.h - the arch-independent
interface that core/mem/fs (and portable drivers) call instead of
poking x86 directly.

API surface:
  - hal_cpu_halt / hal_cpu_pause / hal_cpu_idle
  - hal_intr_enable / disable / save / restore
  - hal_arch_early_init    (one-shot CPU table setup)
  - hal_task_init_stack    (brand-new-task stack pre-build)
  - hal_syscall_exit_to_kernel (SYS_EXIT noreturn epilogue)

x86 backend: kernel/hal/x86/hal_arch.c (single-instruction wrappers
+ delegation to existing gdt/idt/tss init) and
kernel/hal/x86/task_arch.c (the task_trampoline + stack layout
that used to inline in task.c).

Migrations: task.c, lock.c, vc.c, kernel.c, syscall.c lost all
direct __asm__ and their gdt.h/idt.h/tss.h includes. Legacy PC
drivers (pit, ps2) kept their port I/O (PC-only) but switched
their sti+hlt idle to hal_cpu_idle. virtio-blk pause -> hal_cpu_pause.

struct task.esp widened uint32_t -> uintptr_t so the same
context_switch extern works on 32 and 64-bit arches.

Pitfall codified: sti+hlt is a CPU-guaranteed atomic pair (Intel
SDM Vol 2: sti blocks IRQ delivery for ONE instruction boundary).
Splitting into hal_intr_enable + hal_cpu_halt reintroduces the
IRQ-posted-between-the-two race; hence hal_cpu_idle as its own
primitive.

Boot test unchanged: vmm + kmalloc + exFAT + bcache + preempt
(104M hog ticks) + VC + shell all pass.

Deferred (best done together with x64 paging in M20):
  - kernel/mem/vmm.c still pokes CR0/CR3/CR4/invlpg directly.
    hal_map/hal_unmap is the right shape but premature without
    multiple page-table formats in hand.
  - kernel/core/syscall.c still includes idt.h for the x86
    specific struct int_frame.

+++ kernel/includes/task.h
@@ -50,7 +50,7 @@ struct task {
     char     name[TASK_NAME_MAX + 1];
     int      pid;
     enum task_state state;
-    uint32_t esp;                       /* saved kernel-stack pointer */
+    uintptr_t esp;                      /* saved kernel-stack pointer (HAL-typed) */
     void*    kstack_base;               /* kmalloc'd; freed at reap */
     struct task* next;                  /* circular run-queue link */
     /* M14: optional per-task output binding.  When non-NULL, console.c
